Add setup info to prevent possible crashes
You might have some activities in your project that need a theme that is descendant of the Theme.AppCompat theme. When installing react-native-bootsplash, you'll need to specify the theme for these activities, otherwise it will take the theme @style/BootTheme which is not a descendant of Theme.AppCompat and you'll get a crash `Caused by java.lang.IllegalStateException: You need to use a Theme.AppCompat theme (or descendant) with this activity.`. This happens for example with the Google Cast SDK.